test_sentry_sdk/requirements.txt
flask
sentry-sdk
opentelemetry-sdk
opentelemetry-api
opentelemetry-distro
opentelemetry-instrumentation
opentelemetry-semantic-conventions
